requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.2
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-environ==0.13.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyMySQL==1.1.2
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.5.5
text-unidecode==1.3
urllib3==2.6.3
whitenoise==6.11.0
